Move registrationForm component
Move registrationForm component, change input field styled
