+++ frontend/src/Router.jsx
@@ -2,14 +2,17 @@ import { Route, Routes } from 'react-router-dom'
 import Landing from './pages/landing/LandingController'
 import Post from './pages/post/PostPageController'
 
-import Registration from './pages/registration/RegistrationPageController'
+import Registration from './pages/registration/RegistrationPagesController'
 
 const Router = () => (
     <Routes>
         <Route path="/" element={<Landing />} />
         <Route path="/post/:id" element={<Post />} />
 
+
         <Route path="/signUp" element={<Registration />} />
+
+
     </Routes>
 )
